function/requirements.txt
functions-framework==3.4.0
requests==2.31.0
google-cloud-bigquery==3.11.4
google-auth==2.23.4
